+++ mizani/bounds.py
@@ -162,11 +162,18 @@ def rescale_max(x, to=(0, 1), _from=None):
     array([ 0.,  4.,  8., 12., 16., 20.])
 
     If :python:`max(x) < _from[1]` then values will be
-    scaled beyond the requested (:python:`to[1]`) maximum.
+    scaled beyond the requested maximum (:python:`to[1]`).
 
     >>> rescale_max(x, to=(1, 3), _from=(-1, 6))
     array([0., 1., 2., 3., 4., 5.])
 
+    If the values are the same, they taken on the requested maximum.
+    This includes an array of all zeros.
+
+    >>> rescale_max([5, 5, 5])
+    array([1., 1., 1.])
+    >>> rescale_max([0, 0, 0])
+    array([1, 1, 1])
     """
     array_like = True
 
@@ -180,9 +187,14 @@ def rescale_max(x, to=(0, 1), _from=None):
         x = np.asarray(x)
 
     if _from is None:
-        _from = np.array([np.min(x), np.max(x)])
+        _from = (np.min(x), np.max(x))
 
-    out = x/_from[1] * to[1]
+    if np.any(x < 0):
+        out = rescale(x, (0, to[1]), _from)
+    elif np.all(x == 0) and _from[1] == 0:
+        out = np.repeat(to[1], len(x))
+    else:
+        out = x/_from[1] * to[1]
 
     if not array_like:
         out = out[0]